Add iter APIs more suitable for bindings
Add a few public APIs:
- ag_account_get_settings_iter()
- ag_account_service_get_settings_iter()
- ag_account_settings_iter_free()

These methods allocate (and the last one frees) an AgAccountSettingIter
structure on the heap.
Also, they address an issue with the storing of the prefix key: the
C API is assuming that the prefix key remains valid for the whole
duration of the iteration loop, but auto-generated bindings are not
aware of this and free the prefix string as soon as possible.