AgService: load the XML file before returning tags
If the service already existed in the DB and ag_service_get_tags() or
ag_service_has_tag() was the first method called after instantiating a
service, the return value would have been NULL, regardless of the real
tags: this happened because the XML file was not being read at all.

This changes forces the XML to be loaded when the above functions are
invoked.