Account: make code flow more logical
The AgAccountChanges pointer was been set to 0 just before an "if"
statement which could cause the functions to return. It looked like the
AgAccountChanges structure could be leaked in this case, even if in fact
the return statement could be hit only if AgAccountChanges was already
NULL.
However, let's make the code more obvious and robust against future
changes.