Python: restore functionality of Accounts.Manager()
The default implementation for PyGObject constructors is to call
g_object_new(). However, objects which implement the GInitable interface
must be created with g_initable_new() (or alternatively,
g_initable_init() needs to be called right after the object has been
created), but PyGObject doesn't do that for us.
Therefore, we override the Accounts.Manager() constructor to call
ag_manager_new(), which we trust does the right thing.